Fix implementation of Dual Directional Input, Combination Mode None (#220)
* SOCD clean the dual input in Combination Mode None

a later change will apply dualState to the gamepad state properly when
in None mode, but we need to SOCD clean it first (and unlike Combined
mode, it should not be compared to the gamepadState, as None mode should
mean these inputs act independently)

* assign dual inputs to gamepad in Combination Mode None

process() was totally ignoring DUAL_COMBINE_MODE_NONE. this takes the
dual inputs and assigns them to whatever is configured under Dual D-Pad
Mode. this has been SOCD cleaned in aa1f8696

note that this assignment is a direct override when Dual D-Pad Mode is
*different* than the current "normal" D-Pad Mode (as their inputs
contribute to different fields), but if they are the same, the two
inputs contribute in essentially the same as Mixed mode and need to be
SOCD cleaned.