Function ("Fn") button for hotkeys, changing hotkeys themselves via webconfig (#239)
* preliminary version of Fn button functionality

* reimplement Fn as 'aux' rather than gamepad button

same general idea, but we don't reserve a gamepad mask for something
that will never be used as a gamepad button this way. this also
resurrects the "settings pin" idea and the remnants of its aux
mechanism, while still allowing it to be a web configurable pin. it is
still outwardly called the Fn or Function button, but it is assigned to
a separate gamepad state and thus is a bit more obvious that it is *not*
a button that will get sent to the host

* hotkey via web configurator part 1

this removes the "hardcoded" hotkeys and instead allows arbitrary aux,
button, and dpad masks to be associated with the preexisting hotkey
actions. thus, S1 + S2 + Up -> Home can be replaced with Fn + Up -> Home
or whatever else people can concoct.

defaults mappings are retained as well. removing the F1 and F2 mappings
has touched a lot of board configs to remove the 8 existing hotkey
mapping and actions, but they were all identical anyway, so now they're
in the dynamic config rather than defines.

* load hotkeys at gamepad init

hotkeys don't change at runtime so we can just do this to be slightly
more efficient. also, some related code cleanup that I missed.

* only process one hotkey at a time

I feel like firing multiple hotkeys is not what anyone intends to do,
and would only lead to mistakes, and just gets masked by the code
anyway, so avoid it by making this if-else all the way down

* update test getGamepadOptions to match new API

* add Fn and remove hotkeys from Stress board config

* add default defines for hotkey settings

this allows for overrides in BoardConfig.h if so desired, but does not
add any back since they were all the default anyway

* add Fn/remove hotkeys for RP2040MiniBreakoutBoard

* Implement web-config UI for new hotkey settings

* remove deprecated F1 and F2 masks

* basic documentation of the hotkey remapping

* Make hotkey UI better

* update AlpacaOwO board config for Fn + configurable hotkeys

* add hotkey migration to keep old configs on F1/F2 mappings

---------

Co-authored-by: deeebug <77402236+deeebug@users.noreply.github.com>