fix 0% dead zone (#550)
* add scaled radial deadzones

Replace the axial deadzone calculations with scaled radial deadzone as demonstrated here - https://web.archive.org/web/20210827143735/https://www.gamedeveloper.com/disciplines/doing-thumbstick-dead-zones-right

There are a few calculations shared between this and the adjustCircularity function so that was brought into process.

* update magic number

After testing with dead zone sizes 1%-100%, I noticed that 1%-2% requires a larger "magic number". There was not a perceivable difference in scaling at the rest of the dead zone ranges. I felt it better to accommodate the worst case to ensure 100% movement range can always be reached.

* fix 0% dead zone

fix for dead zone if user sets it to 0%