use new A3 and A4 to emulate tpad click position on PS4 (#1060)
* use touchpad + dpad to emulate tpad click position on PS4

when the TP "button" is pressed, we formerly did nothing with the
touchpad data, including not even sending the bit to indicate there is
touchpad data, which was fine for normal purposes. however, we can use
the dpad to emulate position data as if a finger was actually touching
it, and send touchpad data for the eight directions. this will allow us
to simulate clicking the left vs. right sides of the touchpad, which
will satisfy the PS1-on-PS4 emulation of the Start and Select buttons.

this does *not* expose the different position clicks as an action or
anything like that, hoping that this behavior comes with no real
downside, since (AFAIK) touchpad click position data isn't used all that
often on PS4/5 games.

Resolves #924

* change touchpad left/right implementation to actions + hotkeys

rather than doing the implicit dpad direction + A2 = touchpad location
data, define new buttons A3 and A4 across the gamepad, and use them to
be touchpad left and right, respectively. we still need to set the
touchpad data indicators on PS4 mode to trigger this, but now it's an
explicit event rather than implicit thing. both direct pin assignment
and hotkeys have been tested