Add a generic USB HID mode (#1061)
* copy the existing HID driver to a PS3 driver

the intent is to redo the HID driver as a generic, quirkless, expandable
(!), and customizable (!!) driver so that we have a mode that is free of
odd VID/PID dependencies, report layouts, and so on. to accomplish that,
we want to retain the existing HID mode, which has some PS3 quirks, as a
separate driver.

configurations referring to this driver will continue to point at it in
the future, as this is an internal rename and move (i.e. the new HID
driver will be truly new, a different config entry and so on). the PS3
mode was tested in a gamepad tester in Linux, and via SSF2THDR and MvC2
on PS3 --- all seems well, including the analogs on MvC2 and the
functionality of the PS button on both.

* reimplement HID mode as a generic (non-PS-ian) gamepad

this just covers the basics; 16 buttons (the last two don't work yet) +
a hat. TODOs include getting the axes back, and then will begin some
level of customization (be it by defines or protobuf), but that'll
happen later.

I also need to see how some devices (e.g. converters, maybe an Undamned
USB decoder, maybe even an Analogue Pocket Dock) handle this, because
this alone might be usable for those things. maybe.

* enable and define up to 32 buttons for Generic mode

this draws a direct line from GAMEPAD_MASK values to the HID report for
Generic mode, including the dpad Up/Down/Left/Right values as buttons,
and defining 12 "extra" buttons that are only used/output in this mode

* add Generic HID mode to the settings and boot dropdowns

* add analog emulation of L and R sticks to generic HID

* pin mapping labels and things for Extra 1 -- 12

* relabel Extra buttons for specific consoles as "-"

* just a couple trivial cleanups