remove the JS Slider addon, move it to core (#1152)
this removes a bunch of code to support the addon, naturally, but the
idea is the same, just in core. of note, the default mode when no slider
pins are active has been folded into the core gamepad setting, so a tiny
migration copies the user's default mode into the gamepad setting in
order to not violate expectations. the user can still change this
setting there via webconfig or hotkeys.

this also updates the DDI addon to get the active mode (rather than the
settings mode) so that it doesn't do weird cleaning and/or override
things, and updates the display code to show the active mode

Co-authored-by: Luke A <arntsonl@gmail.com>