Hotkeys, Crash, Menu Fix + Tidying (#1324)
* Fix for hotkeys not allowing repeats.

Re-arrange on the main menu screen code.

Added virtual to GPEvent destructors to make sure we are calling all the way up the chain.

* Lock-up fixes, new/delete was happening while the update() function was being called by display

Some code clean-ups, moved all of the gpscreen destructors to virtual so we do not have to switch() the delete function.

Moved some init functions around so we always call init() first

* Add ability to exit menu with pin-based menu toggle. Fix rapid-press of menu toggle