Analog Display Fix for Levers (Haute42 X Analog should work now) (#1541)
* Removed the erroneous map() function from GPGFX_UI and added it to GPLever.cpp

Modified the GPLever analog to use the proper analog values

Analog movement is now reliant on the base radius * analog value / joystick mid

* I don't know why these didn't add....

* Fixed analog.cpp so that the center is not long a conditional with a ceiling, but instead a max(x, 1.f) and using the mid for analog

Also fixed gplever to display analog in the right spot

* Removing map feature as we do not need it

* Fix a few analog mid points, clamping issues on analog.cpp.

Now all calculations for analog should be based on the mid-point. Max if mid is 0x8000 is 0x10000, Max if mid is 0x7FFFF is 0xFFFE. And all of these get converted down to 8-bit for analog values. We do one clamp if analog is 1.f (0x10000) to 0xFFFF